requirements.txt
praw
python-dotenv
kafka-python
json